Added the on startup event

+++ 4/storaging/app/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, String, BigInteger, LargeBinary
+from sqlalchemy import Column, String, Integer, LargeBinary
 
 from .database import Base
 
@@ -6,12 +6,12 @@ from .database import Base
 class Node(Base):
     __tablename__ = "nodes"
 
-    hash = Column(BigInteger, unique=True, primary_key=True)
+    hash = Column(Integer, unique=True, primary_key=True)
     ip_address = Column(String, unique=True)
 
 
 class DataItem(Base):
     __tablename__ = "storage"
 
-    hash = Column(BigInteger, unique=True, primary_key=True)
+    hash = Column(Integer, unique=True, primary_key=True)
     data = Column(LargeBinary)
